issue #1, move code from app to home to get working home page again
